+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 **Implemented enhancements:**
 
 - Added Denmark Holiday Driver
+- New function 'isWorkingDay' added that determines whether a date represents a working day or not. A working day is 
+  considered a date that is neither a holiday nor falls into the weekend.
 - Refactoring and cleanup of unit tests
 
 **Resolved issues:**